Attack Projectiles
Setup the first animation to have real collision! Essentially, during the attacks animation we specify when to spawn the projectile which will fetch the correct projectile for the attack and add it as a child of the weapon. Problems may arise if the weapon is deleted while the projectile is still alive. Possibly use a projectile factory to spawn projectile in a more permanent location.