Weapons System Restructure
Weapons can have up to three abilities loaded from weapon data. Each character, weapon, ability, and projectile keeps track of the damage stats, with weapon and projectile keeping track of d damage data.
Damage stats are the general stats of the weapon: crit, damage mults, effects
Damage data is all data related to damage: damage, damage type, damage stats
When an ability is casted we create the projectile and apply the abilities damage stats. Then, we create a copy of the weapon stats and apply the characters damage stats to it before applying that to the projectile and inserting it into the tree.