Entity Composition
With the direction of the game changing and requirements moving along with that, we have decided to revaluate the process of developing entities.
Through the use of components that each handle a specific task, we can build unique entities from these building blocks and manipulate their stats and behaviours through the editor. The initial setup is substantial, but the pay off is a much faster and more robust method for developing new entities in the future!