Modify damage instancing
Weapons and weapon skill may be loaded up with as many damage instances as you want. A damage instance represents a damage type and a damage value as a percent of the base damage where the base damage is supplied by the caster.
TL;DR a attacks can deal damage of multiple types.