Not the Slimes!
Have implemented a very simple slime enemy. At the moment it moves about in a random direction at random intervals. It can take damage and has a death animation, but I am still in the works of developing a component that lets it track nearby targets and go on the offensive.